requirements.txt
pandas
numpy
pyyaml
python-dotenv
ccxt
requests
ta
scikit-learn
xgboost